exerchk fix (trunk only)
     Post-3.4.3 problem noticed by Michael:  some code which was supposed
to be inside the end of a loop got placed in front of the wrong closing
brace, ending up beyond that loop.  As a result, decay of old attribute
exercise wasn't working as intended, and out of bounds array modification
was occurring (clobbering u.abon.a[0], which probably messed up resetting
strength when taking off rings of gain strength which were worn during an
exercise check).