Guidebook update (paranoid_confirmation 3 of 2) (trunk only)
     This started out just documenting the commands where use of the
new paranoid_confirmation option was relevant, but it end up sprawling
to other stuff so I left it out of the paranoid_confirmation patch.
Eventually I changed all the commands with long-ish descriptions to use
a single line summary of the what the command does, with any additional
explanation or examples forced into a separate paragraph instead of
just being appended to the summary.  It increases the number of lines
and probably pages in the document, but I think it makes skimming over
the list of actual commands much easier.

     A couple of unmodified command descriptions are 'f' and 'Q'.  The
only way I could avoid the temptation to discard "quiver sack" was to
leave those alone entirely.

     A couple of others received some spoiler-ish additions, notably
#offer (which doesn't actually give anything away) and #pray (where
someone might assume that the command is useless if their very first
attempt gets rejected).  I also added tips for two-weapon combat (how
to set up to use it, not when or why to use it) that ended up being much
more verbose than planned.

     I don't know whether nroff+tmac.n offers a better way to get a
non-indented paragraph than using a labeled paragraph with an empty
label; .lp "" achieved what I wanted so I used it quite a bit.  I also
wanted the value lists for number_pad and paranoid_confirmation to not
be indented but failed to figure out how to do that properly.  In
Guidebook.mn they're still indented; in Guidebook.tex number_pad fakes
it using fixed-with tt font, paranoid_confirmation approximates it with
a ridiculous indentation hack.  The number_pad result is wrong, but I've
given up.  "~0" lines up with "-1", but "~1" through "~4" line up with
the minus sign instead of with the 1 as if that unbreakable space prefix
wasn't there.