tty extmenu option
     When the 'extmenu' boolean was set, typing '#' in wizard mode triggered
a crash.  The total number of extended commands exceeded the capacity of
extcmd_via_menu() and the code it had to check for that was only in effect
if DEBUG was defined.  This boosts the limit from 40 (actally 38) to 50 (48)
and enables the checking code for BETA as well as for DEBUG.

     Almost all of this diff is indentation.